# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Auto Forex Trading API
  version: 1.0.0
  description:
    API for managing algorithmic forex trading operations across multiple
    OANDA accounts
  contact:
    name: Auto Forex Support
    email: support@autoforex.example.com
  license:
    name: Proprietary
paths:
...
  /api/market/accounts/:
    get:
      operationId: list_oanda_accounts
      description: Retrieve all OANDA accounts for the authenticated user
      summary: GET /api/market/accounts/
      tags:
        - Market - Accounts
      security:
        - jwtAuth: []
        - cookieAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OandaAccounts"
          description: List of OANDA accounts
        "401":
          description: Authentication required
    post:
      operationId: create_oanda_account
      description: Add a new OANDA account for the authenticated user
      summary: POST /api/market/accounts/
      tags:
        - Market - Accounts
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OandaAccountsRequest"
        required: true
      security:
        - jwtAuth: []
        - cookieAuth: []
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OandaAccounts"
          description: Account created successfully
        "400":
          description: Invalid data
        "401":
          description: Authentication required
  /api/market/accounts/{account_id}/:
    get:
      operationId: get_oanda_account_detail
      description:
        Retrieve detailed information about a specific OANDA account including
        live data from OANDA API
      summary: GET /api/market/accounts/{account_id}/
      parameters:
        - in: path
          name: account_id
          schema:
            type: integer
          description: OANDA account database ID
          required: true
      tags:
        - Market - Accounts
      security:
        - jwtAuth: []
        - cookieAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OandaAccounts"
          description: Account details retrieved successfully
        "401":
          description: Authentication required
        "404":
          description: Account not found
    put:
      operationId: update_oanda_account
      description: Update a specific OANDA account
      summary: PUT /api/market/accounts/{account_id}/
      parameters:
        - in: path
          name: account_id
          schema:
            type: integer
          description: OANDA account database ID
          required: true
      tags:
        - Market - Accounts
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OandaAccountsRequest"
        required: true
      security:
        - jwtAuth: []
        - cookieAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OandaAccounts"
          description: Account updated successfully
        "400":
          description: Invalid data
        "401":
          description: Authentication required
        "404":
          description: Account not found
    delete:
      operationId: delete_oanda_account
      description: Delete a specific OANDA account
      summary: DELETE /api/market/accounts/{account_id}/
      parameters:
        - in: path
          name: account_id
          schema:
            type: integer
          description: OANDA account database ID
          required: true
      tags:
        - Market - Accounts
      security:
        - jwtAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Account deleted successfully
        "400":
          description: Account is in use
        "401":
          description: Authentication required
        "404":
          description: Account not found
...
components:
  schemas:
    ApiTypeEnum:
      enum:
        - practice
        - live
      type: string
      description: |-
        * `practice` - Practice
        * `live` - Live
...
    JurisdictionEnum:
      enum:
        - US
        - JP
        - OTHER
      type: string
      description: |-
        * `US` - United States
        * `JP` - Japan
        * `OTHER` - Other/International
...
    OandaAccounts:
      type: object
      description: Serializer for OANDA account.
      properties:
        id:
          type: integer
          readOnly: true
        account_id:
          type: string
          description: OANDA account ID
          maxLength: 100
        api_type:
          allOf:
            - $ref: "#/components/schemas/ApiTypeEnum"
          description: |-
            API endpoint type (practice or live)

            * `practice` - Practice
            * `live` - Live
        jurisdiction:
          allOf:
            - $ref: "#/components/schemas/JurisdictionEnum"
          description: |-
            Regulatory jurisdiction for this account

            * `US` - United States
            * `JP` - Japan
            * `OTHER` - Other/International
        currency:
          type: string
          description: Account base currency
          maxLength: 3
        balance:
          type: string
          format: decimal
          pattern: ^-?\d{0,13}(?:\.\d{0,2})?$
          readOnly: true
          description: Current account balance
        margin_used:
          type: string
          format: decimal
          pattern: ^-?\d{0,13}(?:\.\d{0,2})?$
          readOnly: true
          description: Margin currently used by open positions
        margin_available:
          type: string
          format: decimal
          pattern: ^-?\d{0,13}(?:\.\d{0,2})?$
          readOnly: true
          description: Margin available for new positions
        unrealized_pnl:
          type: string
          format: decimal
          pattern: ^-?\d{0,13}(?:\.\d{0,2})?$
          readOnly: true
          description: Unrealized profit/loss from open positions
        is_active:
          type: boolean
          description: Whether the account is active
        is_default:
          type: boolean
          description: Whether this is the default account for market data collection
        created_at:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the account was added
        updated_at:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the account was last updated
      required:
        - account_id
    OandaAccountsRequest:
      type: object
      description: Serializer for OANDA account.
      properties:
        account_id:
          type: string
          minLength: 1
          description: OANDA account ID
          maxLength: 100
        api_token:
          type: string
          writeOnly: true
          minLength: 1
          description: OANDA API token (will be encrypted)
        api_type:
          allOf:
            - $ref: "#/components/schemas/ApiTypeEnum"
          description: |-
            API endpoint type (practice or live)

            * `practice` - Practice
            * `live` - Live
        jurisdiction:
          allOf:
            - $ref: "#/components/schemas/JurisdictionEnum"
          description: |-
            Regulatory jurisdiction for this account

            * `US` - United States
            * `JP` - Japan
            * `OTHER` - Other/International
        currency:
          type: string
          minLength: 1
          description: Account base currency
          maxLength: 3
        is_active:
          type: boolean
          description: Whether the account is active
        is_default:
          type: boolean
          description: Whether this is the default account for market data collection
      required:
        - account_id
        - api_token
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.autoforex.example.com
    description: Production server
...